+++ web/api/router.go
@@ -29,12 +29,12 @@ type paramSearch struct {
 }
 
 type paramId struct {
-	Id int `uri:"id"`
+	Id int64 `uri:"id"`
 }
 
 type paramId2 struct {
-	Id  int `uri:"id"`
-	Id2 int `uri:"id2"`
+	Id  int64 `uri:"id"`
+	Id2 int64 `uri:"id2"`
 }
 
 func mustLogin(c *gin.Context) {
@@ -61,9 +61,9 @@ func RegisterRoutes(app *gin.RouterGroup) {
 	mod := reflect.TypeOf(model.Tunnel{})
 	//app.POST("/project/:id/tunnels", curdApiListById(mod, "project_id"))
 	app.POST("/tunnels", curdApiList(mod))
-	app.POST("/tunnel", curdApiCreate(mod, nil, nil))       //TODO 启动
-	app.DELETE("/tunnel/:id", curdApiDelete(mod, nil, nil)) //TODO 停止
-	app.PUT("/tunnel/:id", curdApiModify(mod, []string{""},nil, nil))    //TODO 重新启动
+	app.POST("/tunnel", curdApiCreate(mod, nil, nil))                  //TODO 启动
+	app.DELETE("/tunnel/:id", curdApiDelete(mod, nil, nil))            //TODO 停止
+	app.PUT("/tunnel/:id", curdApiModify(mod, []string{""}, nil, nil)) //TODO 重新启动
 	app.GET("/tunnel/:id", curdApiGet(mod))
 
 	app.GET("/tunnel/:id/start", tunnelStart)
@@ -76,7 +76,7 @@ func RegisterRoutes(app *gin.RouterGroup) {
 	//app.POST("/tunnel/:id/links", curdApiListById(mod, "tunnel_id"))
 	app.POST("/links", curdApiList(mod))
 	app.DELETE("/link/:id", curdApiDelete(mod, nil, nil)) //TODO 停止
-	app.PUT("/link/:id", curdApiModify(mod, []string{""},nil, nil))
+	app.PUT("/link/:id", curdApiModify(mod, []string{""}, nil, nil))
 	app.GET("/link/:id", curdApiGet(mod))
 
 	//设备管理
@@ -85,7 +85,7 @@ func RegisterRoutes(app *gin.RouterGroup) {
 	app.POST("/devices", curdApiList(mod))
 	app.POST("/device", curdApiCreate(mod, nil, nil))
 	app.DELETE("/device/:id", curdApiDelete(mod, nil, nil))
-	app.PUT("/device/:id", curdApiModify(mod, []string{""},nil, nil))
+	app.PUT("/device/:id", curdApiModify(mod, []string{""}, nil, nil))
 	app.GET("/device/:id", curdApiGet(mod))
 
 	//插件管理
@@ -93,15 +93,15 @@ func RegisterRoutes(app *gin.RouterGroup) {
 	app.POST("/plugins", curdApiList(mod))
 	app.POST("/plugin", curdApiCreate(mod, nil, nil))
 	app.DELETE("/plugin/:id", curdApiDelete(mod, nil, nil))
-	app.PUT("/plugin/:id", curdApiModify(mod, []string{""},nil, nil))
+	app.PUT("/plugin/:id", curdApiModify(mod, []string{""}, nil, nil))
 	app.GET("/plugin/:id", curdApiGet(mod))
 
 	//项目管理
 	mod = reflect.TypeOf(model.Project{})
 	app.POST("/projects", curdApiList(mod))
-	app.POST("/project", curdApiCreate(mod,  projectBeforeCreate, projectAfterCreate))
-	app.DELETE("/project/:id", curdApiDelete(mod, nil,projectAfterDelete))
-	app.PUT("/project/:id", curdApiModify(mod, []string{""},nil, projectAfterModify))
+	app.POST("/project", curdApiCreate(mod, projectBeforeCreate, projectAfterCreate))
+	app.DELETE("/project/:id", curdApiDelete(mod, nil, projectAfterDelete))
+	app.PUT("/project/:id", curdApiModify(mod, []string{""}, nil, projectAfterModify))
 	app.GET("/project/:id", curdApiGet(mod))
 
 	app.POST("/project-import", projectImport)
@@ -111,12 +111,11 @@ func RegisterRoutes(app *gin.RouterGroup) {
 	//元件管理
 	mod = reflect.TypeOf(model.Element{})
 	app.POST("/elements", curdApiList(mod))
-	app.POST("/element", curdApiCreate(mod, elementBeforeCreate,nil))
+	app.POST("/element", curdApiCreate(mod, elementBeforeCreate, nil))
 	app.DELETE("/element/:id", curdApiDelete(mod, elementBeforeDelete, nil))
-	app.PUT("/element/:id", curdApiModify(mod, []string{""},nil,nil))
+	app.PUT("/element/:id", curdApiModify(mod, []string{""}, nil, nil))
 	app.GET("/element/:id", curdApiGet(mod))
 }
-
 
 func replyList(ctx *gin.Context, data interface{}, total int64) {
 	ctx.JSON(http.StatusOK, gin.H{"data": data, "total": total})
